refractor: move font functions -> fontUtility, fontData.ts
