+++ frontend/src/data/fontData.ts
@@ -1,0 +1,197 @@
+export const defaultFontsData = [
+    {
+      fallback: true,
+      label: "notoSerifJP-regular",
+      url: "/fonts/NotoSerifJP-Regular.otf",
+    },
+    {
+      fallback: false,
+      label: "notoSerifJP-regular-bold",
+      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "notoSerifJP-regular-italic",
+      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "basilia-bold",
+      url: "/fonts/Basilia-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "basilia-italic",
+      url: "/fonts/Basilia-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "basilia",
+      url: "/fonts/Basilia-Reg.woff",
+    },
+    {
+      fallback: false,
+      label: "futura-bold",
+      url: "/fonts/Futura-Dem-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "futura-italic",
+      url: "/fonts/Futura-Dem-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "futura",
+      url: "/fonts/Futura-Dem.woff",
+    },
+    {
+      fallback: false,
+      label: "montserrat",
+      url: "/fonts/Montserrat-Regular.ttf",
+    },
+    {
+      fallback: false,
+      label: "montserrat-bold",
+      url: "/fonts/Montserrat-Bold.ttf",
+    },
+    {
+      fallback: false,
+      label: "montserrat-italic",
+      url: "/fonts/Montserrat-Italic.ttf",
+    },
+    {
+      fallback: false,
+      label: "museosans",
+      url: "/fonts/MuseoSans_500.woff",
+    },
+    {
+      fallback: false,
+      label: "museosans-bold",
+      url: "/fonts/MuseoSans_500-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "museosans-italic",
+      url: "/fonts/MuseoSans_500-Italic.otf",
+    },
+    {
+      fallback: false,
+      label: "ttcorals",
+      url: "/fonts/TT-Corals-Regular.woff",
+    },
+    {
+      fallback: false,
+      label: "ttcorals-bold",
+      url: "/fonts/TT-Corals-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "ttcorals-italic",
+      url: "/fonts/TT-Corals-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "bison",
+      url: "/fonts/Basilia-Reg.woff",
+    },
+    {
+      fallback: false,
+      label: "bison-bold",
+      url: "/fonts/Bison-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "bison-italic",
+      url: "/fonts/Basilia-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "arial",
+      url: "https://fonts.cdnfonts.com/s/29105/ARIAL.woff",
+    },
+    {
+      fallback: false,
+      label: "arial-bold",
+      url: "https://fonts.cdnfonts.com/s/29105/ARIALBD.woff",
+    },
+    {
+      fallback: false,
+      label: "arial-italic",
+      url: "https://fonts.cdnfonts.com/s/29105/ARIALI.woff",
+    },
+    {
+      fallback: false,
+      label: "overpass",
+      url: "https://fonts.cdnfonts.com/s/12274/Overpass_Regular.woff",
+    },
+    {
+      fallback: false,
+      label: "overpass-bold",
+      url: "https://fonts.cdnfonts.com/s/12274/Overpass_Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "overpass-italic",
+      url: "https://fonts.cdnfonts.com/s/12274/overpassitalic.woff",
+    },
+    {
+      fallback: false,
+      label: "roboto",
+      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Regular.woff",
+    },
+    {
+      fallback: false,
+      label: "roboto-bold",
+      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "roboto-italic",
+      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "rubik",
+      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Regular.woff",
+    },
+    {
+      fallback: false,
+      label: "rubik-bold",
+      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Bold.woff",
+    },
+    {
+      fallback: false,
+      label: "rubik-italic",
+      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Italic.woff",
+    },
+    {
+      fallback: false,
+      label: "brcobane",
+      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-Regular-BF654d96a1718fa.woff",
+    },
+    {
+      fallback: false,
+      label: "brcobane-bold",
+      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-Bold-BF654d96a1ac1b0.woff",
+    },
+    {
+      fallback: false,
+      label: "brcobane-italic",
+      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-RegularItalic-BF654d96a1a8637.woff",
+    },
+    {
+      fallback: false,
+      label: "calibri",
+      url: "https://fonts.gstatic.com/l/font?kit=J7afnpV-BGlaFfdAhLEY6w&skey=a1029226f80653a8&v=v15",
+    },
+    {
+      fallback: false,
+      label: "calibri-bold",
+      url: "https://fonts.gstatic.com/l/font?kit=J7aanpV-BGlaFfdAjAo9_pxqHw&skey=cd2dd6afe6bf0eb2&v=v15",
+    },
+    {
+      fallback: false,
+      label: "calibri-italic",
+      url: "https://fonts.gstatic.com/l/font?kit=J7adnpV-BGlaFfdAhLQo6btP&skey=36a3d5758e0e2f58&v=v15",
+    },
+  ];
