+++ frontend/src/pages/Verifcation/VerificationPageBackup.tsx
@@ -5,13 +5,14 @@ import { BootcampResponse, StudentResponse, TemplateResponse } from "../../util/
 import { SpinnerDefault } from "../../components/MenuItems/Loaders/SpinnerDefault";
 import { useEffect, useRef, useState } from "react";
 import { Form, Viewer } from "@pdfme/ui";
-import { generatePDF, getFontsData, getPlugins } from "../../util/helper";
+import { generatePDF, getPlugins } from "../../util/helper";
 import { makeTemplateInput } from "../../templates/baseTemplate";
 import { mapTemplateInputsToTemplateViewerSingle, templateInputsSingleBootcampandTemplate } from "../../util/dataHelpers";
 import { PublishButton } from '../../components/MenuItems/Buttons/PublishButton';
 import { SuccessIcon } from '../../components/MenuItems/Icons/SuccessIcon';
 import { NextIcon } from '../../components/MenuItems/Icons/NextIcon';
 import logoBlack from '/icons/logoBlack.png'
+import { getFontsData } from '../../util/fontsUtility';
 
 type Props = {
     getStudentByVerificationCode: (verificationCode: string) => void;
