+++ frontend/src/util/helper.ts
@@ -9,15 +9,15 @@ import { useLoadingMessage } from "../components/Contexts/LoadingMessageContext"
 import JSZip from 'jszip';
 import { saveAs } from 'file-saver';
 import * as pdfjsLib from 'pdfjs-dist';
-import { updateStudentPreviewImage } from "../services/bootcampService";
 import { initApiEndpoints } from "../services/apiFactory";
+import { getFontsData } from "./fontsUtility";
+
 
 export const getToken = (): string => {
   const jwtToken = document.cookie
-    .split('; ')
-    .find(c => c.includes('__session'))
-    ?.split('=')[1] || '';
-  // console.log(jwtToken)
+      .split('; ')
+      .find(c => c.includes('__session'))
+      ?.split('=')[1] || '';
   return jwtToken
 }
 
@@ -25,297 +25,6 @@ let api = initApiEndpoints({
   endpointUrl: import.meta.env.VITE_API_URL,
   token: getToken()
 });
-
-
-const fontCache = new Map<string, { label: string; url: string; data: ArrayBuffer }>();
-let userFonts: UserFontResponseDto[]
-let userfontObjList = []
-
-const fetchUserFonts = async () => {
-  userFonts = await api.getUserFonts();
-  for (let i = 0; i < userFonts.length; i++) {
-    const newFont = userFonts[i];
-    userfontObjList.push({
-      fallback: false,
-      label: newFont.fileName,
-      url: newFont.fileUrl,
-    })
-  }
-}
-
-// await fetchUserFonts();
-
-let defaultFontObjList;
-let allFontsList;
-
-(async () => {
-  await fetchUserFonts();
-
-  defaultFontObjList = [
-    {
-      fallback: true,
-      label: "notoSerifJP-regular",
-      url: "/fonts/NotoSerifJP-Regular.otf",
-    },
-    {
-      fallback: false,
-      label: "notoSerifJP-regular-bold",
-      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "notoSerifJP-regular-italic",
-      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "basilia-bold",
-      url: "/fonts/Basilia-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "basilia-italic",
-      url: "/fonts/Basilia-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "basilia",
-      url: "/fonts/Basilia-Reg.woff",
-    },
-    {
-      fallback: false,
-      label: "futura-bold",
-      url: "/fonts/Futura-Dem-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "futura-italic",
-      url: "/fonts/Futura-Dem-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "futura",
-      url: "/fonts/Futura-Dem.woff",
-    },
-    {
-      fallback: false,
-      label: "montserrat",
-      url: "/fonts/Montserrat-Regular.ttf",
-    },
-    {
-      fallback: false,
-      label: "montserrat-bold",
-      url: "/fonts/Montserrat-Bold.ttf",
-    },
-    {
-      fallback: false,
-      label: "montserrat-italic",
-      url: "/fonts/Montserrat-Italic.ttf",
-    },
-    {
-      fallback: false,
-      label: "museosans",
-      url: "/fonts/MuseoSans_500.woff",
-    },
-    {
-      fallback: false,
-      label: "museosans-bold",
-      url: "/fonts/MuseoSans_500-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "museosans-italic",
-      url: "/fonts/MuseoSans_500-Italic.otf",
-    },
-    {
-      fallback: false,
-      label: "ttcorals",
-      url: "/fonts/TT-Corals-Regular.woff",
-    },
-    {
-      fallback: false,
-      label: "ttcorals-bold",
-      url: "/fonts/TT-Corals-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "ttcorals-italic",
-      url: "/fonts/TT-Corals-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "bison",
-      url: "/fonts/Basilia-Reg.woff",
-    },
-    {
-      fallback: false,
-      label: "bison-bold",
-      url: "/fonts/Bison-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "bison-italic",
-      url: "/fonts/Basilia-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "arial",
-      url: "https://fonts.cdnfonts.com/s/29105/ARIAL.woff",
-    },
-    {
-      fallback: false,
-      label: "arial-bold",
-      url: "https://fonts.cdnfonts.com/s/29105/ARIALBD.woff",
-    },
-    {
-      fallback: false,
-      label: "arial-italic",
-      url: "https://fonts.cdnfonts.com/s/29105/ARIALI.woff",
-    },
-    {
-      fallback: false,
-      label: "overpass",
-      url: "https://fonts.cdnfonts.com/s/12274/Overpass_Regular.woff",
-    },
-    {
-      fallback: false,
-      label: "overpass-bold",
-      url: "https://fonts.cdnfonts.com/s/12274/Overpass_Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "overpass-italic",
-      url: "https://fonts.cdnfonts.com/s/12274/overpassitalic.woff",
-    },
-    {
-      fallback: false,
-      label: "roboto",
-      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Regular.woff",
-    },
-    {
-      fallback: false,
-      label: "roboto-bold",
-      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "roboto-italic",
-      url: "https://fonts.cdnfonts.com/s/12165/Roboto-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "rubik",
-      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Regular.woff",
-    },
-    {
-      fallback: false,
-      label: "rubik-bold",
-      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Bold.woff",
-    },
-    {
-      fallback: false,
-      label: "rubik-italic",
-      url: "https://fonts.cdnfonts.com/s/15684/Rubik-Italic.woff",
-    },
-    {
-      fallback: false,
-      label: "brcobane",
-      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-Regular-BF654d96a1718fa.woff",
-    },
-    {
-      fallback: false,
-      label: "brcobane-bold",
-      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-Bold-BF654d96a1ac1b0.woff",
-    },
-    {
-      fallback: false,
-      label: "brcobane-italic",
-      url: "https://fonts.cdnfonts.com/s/107551/BRCobane-RegularItalic-BF654d96a1a8637.woff",
-    },
-    {
-      fallback: false,
-      label: "calibri",
-      url: "https://fonts.gstatic.com/l/font?kit=J7afnpV-BGlaFfdAhLEY6w&skey=a1029226f80653a8&v=v15",
-    },
-    {
-      fallback: false,
-      label: "calibri-bold",
-      url: "https://fonts.gstatic.com/l/font?kit=J7aanpV-BGlaFfdAjAo9_pxqHw&skey=cd2dd6afe6bf0eb2&v=v15",
-    },
-    {
-      fallback: false,
-      label: "calibri-italic",
-      url: "https://fonts.gstatic.com/l/font?kit=J7adnpV-BGlaFfdAhLQo6btP&skey=36a3d5758e0e2f58&v=v15",
-    },
-  ];
-
-  allFontsList = [...defaultFontObjList, ...userfontObjList];
-
-  await (logFontMimeTypes);
-
-  await getFontsData();
-})();
-
-
-
-export const logFontMimeTypes = async () => {
-  await Promise.all(
-    allFontsList.map(async (font) => {
-      try {
-        const response = await fetch(font.url, { method: 'HEAD' });
-        const contentType = response.headers.get('Content-Type');
-
-        console.log(`Font: ${font.label}, MIME type: ${contentType}`);
-      } catch (error) {
-        console.error(`Failed to fetch MIME type for ${font.label}:`, error);
-      }
-    })
-  );
-};
-
-
-export const getFontsData = async () => {
-  const fontDataList = await Promise.all(
-    allFontsList.map(async (font) => {
-      if (!fontCache.has(font.label)) {
-        let data = await getFontFromIndexedDB(font.label);
-        if (!data) {
-          data = await fetch(font.url).then((res) => res.arrayBuffer());
-          await storeFontInIndexedDB(font.label, data);
-        }
-        fontCache.set(font.label, { ...font, data });
-      }
-      return fontCache.get(font.label);
-    })
-  );
-
-  return fontDataList.reduce(
-    (acc, font) => ({ ...acc, [font.label]: font }),
-    {} as Font
-  );
-};
-
-export const refreshUserFonts = async () => {
-  userfontObjList = [];
-
-  const refreshedUserFonts: UserFontResponseDto[] = await api.getUserFonts();
-
-  refreshedUserFonts.forEach((newFont) => {
-      userfontObjList.push({
-          fallback: false,
-          label: newFont.fileName,
-          url: newFont.fileUrl,
-      });
-  });
-
-  allFontsList = [...defaultFontObjList, ...userfontObjList];
-
-  console.log(allFontsList);
-
-  fontCache.clear();
-
-  await getFontsData();
-};
 
 export const readFile = (
   file: File | null,
@@ -733,60 +442,6 @@ export function mapBootcampToSaltData2(TrackName: string, bootcamp: BootcampResp
       displayName: "Fullstack " + TrackName 
   };
 }
-
-const openDB = () => {
-  return new Promise<IDBDatabase>((resolve, reject) => {
-    const request = indexedDB.open('fontDatabase', 1);
-
-    request.onupgradeneeded = (event: IDBVersionChangeEvent) => {
-      const db = (event.target as IDBOpenDBRequest).result;
-      db.createObjectStore('fonts');
-    };
-
-    request.onsuccess = (event: Event) => {
-      resolve((event.target as IDBOpenDBRequest).result);
-    };
-
-    request.onerror = (event: Event) => {
-      reject((event.target as IDBOpenDBRequest).error);
-    };
-  });
-};
-
-const storeFontInIndexedDB = async (label: string, fontData: ArrayBuffer) => {
-  const db = await openDB();
-  return new Promise<void>((resolve, reject) => {
-    const transaction = db.transaction(['fonts'], 'readwrite');
-    const store = transaction.objectStore('fonts');
-    const request = store.put(fontData, label);
-
-    request.onsuccess = () => {
-      resolve();
-    };
-
-    request.onerror = (event: Event) => {
-      reject((event.target as IDBRequest).error);
-    };
-  });
-};
-
-const getFontFromIndexedDB = async (label: string): Promise<ArrayBuffer | null> => {
-  const db = await openDB();
-  return new Promise<ArrayBuffer | null>((resolve, reject) => {
-    const transaction = db.transaction(['fonts'], 'readonly');
-    const store = transaction.objectStore('fonts');
-    const request = store.get(label);
-
-    request.onsuccess = (event: Event) => {
-      resolve((event.target as IDBRequest).result as ArrayBuffer | null);
-    };
-
-    request.onerror = (event: Event) => {
-      reject((event.target as IDBRequest).error);
-    };
-  });
-};
-
 
 export const generateVerificationCode = (tracks: TrackResponse[]): string => {
